Add support for icons within badges
This commit adds a new attribute - `icon` - to `paper-badge`
elements that facilitate the placement of `iron-icon` elements within
badges. This allows clients to easily use icons without the need for
an additional element. According to the designers on the Material
Design team, icons within badges is a use case that should be
supported.